Fix polynomial set degree of enriched
PolynomialSet.get_degree() is merely a label and is, in fact, never
used. Nevertheless it should be probably a minimum for compound
elements. What is the degree of expansion set for P1+B3 on triangle? Is
it 1 or 3?

More correct would be maybe to assign degree 0 to B3 and then take
degree(P1+B3) = max(deg(P1)+deg(B3)) = 1.