add comments about MixedElement nodes
MixedElement nodes are the same as they were in the FFC version,
however, they clearly have a shape mismatch. When, for example, mixing
a DG element and a RT element, one gets a mixed set of
nodes (functionals), some of which map from 2D vector-valued
functions (from the RT element) and some map from scalar-valued
functions (from the DG element).

What should happen, is that all nodes (functionals) map from "3D
vector-valued" functions, the ones from RT ignore the component that
corresponds to the DG field, while the ones from DG ignore the
components that correspond to the RT field.

Or, for example, a mixed CG x CG element will contain the same list of
nodes twice, with no information about which nodes corresponds to
which component in the vector-CG element.  This is then all hacked
around with offsets in ffc/representation.py (see functions
_generate_reference_offsets and _generate_physical_offsets).