Textbook functionals for H(div/curl) elements
Although the existing FIAT functionals are fine for tabulation and
hence normal finite element assembly, they under-integrate for the
purposes of interpolation. This results in sub-optimal convergence in
the H(div/curl) norm of the interpolation error, which should be (for
example)

   ||u - I_q u||_H(div, T) \le C h^q |u|_H^{q+1}(T)

for RTq elements.

To fix this, implement the dual basis using "textbook" integral moment
functionals (even on edges).

These functionals are selected by creating the element with an
appropriate variant string on construction.